src/requirements.txt
alabaster==0.7.9
amqp==2.1.4
appdirs==1.4.3
appnope==0.1.0
Babel==2.3.4
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
bleach==2.0.0
cachetools==2.0.0
celery==4.0.2
certifi==2017.4.17
cffi==1.8.3
configparser==3.5.0
cryptography==1.5.2
cycler==0.10.0
decorator==4.0.11
dill==0.2.6
Django==1.8.18
django-celery==3.2.1
django-crispy-forms==1.6.1
django-extensions==1.7.5
django-registration-redux==1.4
django-widget-tweaks==1.4.1
docutils==0.12
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
gapic-google-cloud-speech-v1==0.15.3
gapic-google-cloud-speech-v1beta1==0.15.3
gapic-google-cloud-vision-v1==0.90.3
google-api-python-client==1.5.5
google-auth==1.0.0
google-auth-httplib2==0.0.2
google-cloud==0.25.0
google-cloud-bigquery==0.24.0
google-cloud-bigtable==0.24.0
google-cloud-core==0.24.1
google-cloud-datastore==1.0.0
google-cloud-dns==0.24.0
google-cloud-error-reporting==0.24.2
google-cloud-language==0.24.1
google-cloud-logging==1.0.0
google-cloud-monitoring==0.24.0
google-cloud-pubsub==0.25.0
google-cloud-resource-manager==0.24.0
google-cloud-runtimeconfig==0.24.0
google-cloud-spanner==0.24.1
google-cloud-speech==0.24.0
google-cloud-storage==1.1.1
google-cloud-translate==0.24.0
google-cloud-vision==0.24.0
google-gax==0.15.11
google-resumable-media==0.0.2
googleapis-common-protos==1.5.2
grpc-google-cloud-speech-v1beta1==0.14.0
grpc-google-iam-v1==0.11.1
grpcio==1.3.0
gunicorn==19.7.1
html5lib==0.999999999
httplib2==0.10.3
idna==2.1
imagesize==0.7.1
ipaddress==1.0.17
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.8
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
kombu==4.0.2
MarkupSafe==0.23
matplotlib==2.0.2
mistune==0.7.4
mpmath==0.19
nbconvert==5.1.1
nbformat==4.3.0
ndg-httpsclient==0.4.2
nose==1.3.7
notebook==5.0.0
numpy==1.12.1
oauth2client==3.0.0
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
ply==3.8
prompt-toolkit==1.0.14
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
proto-google-cloud-speech-v1==0.15.3
proto-google-cloud-speech-v1beta1==0.15.3
proto-google-cloud-vision-v1==0.90.3
protobuf==3.3.0
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
pydot==1.2.3
Pygments==2.1.3
pyOpenSSL==16.2.0
pyparsing==2.2.0
PySoundFile==0.9.0.post1
python-dateutil==2.6.0
pytz==2016.7
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.13.0
rsa==3.4.2
scandir==1.5
scipy==0.19.0
simplegeneric==0.8.1
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sounddevice==0.3.7
SpeechRecognition==3.4.6
subprocess32==3.2.7
sympy==1.0
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
uritemplate==3.0.0
vine==1.1.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
wit==4.2.0
